# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen User & Task API",
        "description": "Dokumentasi API untuk aplikasi manajemen user dan tugas.",
        "contact": {
            "name": "Support",
            "email": "support@yourdomain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Ambil daftar tugas",
                "description": "Display a listing of the resource.\nGET /tasks\nFilter berdasarkan role:\n- Admin: semua task\n- Manager: task yang dibuatnya, ditugaskan padanya, atau ditugaskan ke staff\n- Staff: task yang dibuatnya atau ditugaskan padanya",
                "operationId": "7786da811325606208249ebc8d393d9a",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar tugas"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Tambah tugas baru",
                "description": "Store a newly created resource in storage.\nPOST /tasks\nAkses: Semua role bisa membuat. Manager hanya bisa assign ke staff.",
                "operationId": "a737e990cef5769592bc41ee6a62f7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tugas Baru"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi tugas"
                                    },
                                    "assigned_to_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid-user"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "done"
                                        ],
                                        "example": "pending"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tugas berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{task}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Ambil detail tugas",
                "description": "Display the specified resource.\nGET /tasks/{task}",
                "operationId": "2507ad5dcbf3084e44634aa11b15da4e",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail tugas berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Tugas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update tugas",
                "description": "Update the specified resource in storage.\nPUT /tasks/{task}",
                "operationId": "9b57b01aef11702efbff465e97dd2230",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tugas Update"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi update"
                                    },
                                    "assigned_to_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "uuid-user"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "done"
                                        ],
                                        "example": "in_progress"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-07-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tugas berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Tugas tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Hapus tugas",
                "description": "Remove the specified resource from storage.\nDELETE /tasks/{task}",
                "operationId": "70a8dde6171ddd9371617d15b1b1b750",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tugas berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Tugas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/export/csv": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Export daftar tugas ke CSV",
                "description": "Export daftar tugas ke CSV (Bonus)\nGET /tasks/export/csv\nAkses: semua role (admin, manager, staff)",
                "operationId": "7d334df10a17d13a1d09dba7f3b02c8f",
                "responses": {
                    "200": {
                        "description": "File CSV berhasil diunduh",
                        "content": {
                            "text/csv": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}